// wren-ui/src/types/api.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DialectSQL: any;
  JSON: any;
};
// ...
export type AdjustmentTask = {
  __typename?: 'AdjustmentTask';
  error?: Maybe<Error>;
  invalidSql?: Maybe<Scalars['String']>;
  queryId?: Maybe<Scalars['String']>;
  sql?: Maybe<Scalars['String']>;
  status?: Maybe<AskingTaskStatus>;
  traceId?: Maybe<Scalars['String']>;
};
// ...
export type AskingTask = {
  __typename?: 'AskingTask';
  candidates: Array<ResultCandidate>;
  error?: Maybe<Error>;
  intentReasoning?: Maybe<Scalars['String']>;
  invalidSql?: Maybe<Scalars['String']>;
  queryId?: Maybe<Scalars['String']>;
  rephrasedQuestion?: Maybe<Scalars['String']>;
  retrievedTables?: Maybe<Array<Scalars['String']>>;
  sqlGenerationReasoning?: Maybe<Scalars['String']>;
  status: AskingTaskStatus;
  traceId?: Maybe<Scalars['String']>;
  type?: Maybe<AskingTaskType>;
};

export type AskingTaskInput = {
  knowledgeBaseIds?: InputMaybe<Array<Scalars['String']>>;
  question: Scalars['String'];
  selectedSkillIds?: InputMaybe<Array<Scalars['String']>>;
  threadId?: InputMaybe<Scalars['Int']>;
};

export enum AskingTaskStatus {
  CORRECTING = 'CORRECTING',
  FAILED = 'FAILED',
  FINISHED = 'FINISHED',
  GENERATING = 'GENERATING',
  PLANNING = 'PLANNING',
  SEARCHING = 'SEARCHING',
  STOPPED = 'STOPPED',
  UNDERSTANDING = 'UNDERSTANDING',
}

export enum AskingTaskType {
  GENERAL = 'GENERAL',
  MISLEADING_QUERY = 'MISLEADING_QUERY',
  TEXT_TO_SQL = 'TEXT_TO_SQL',
}
// ...
export enum ChartTaskStatus {
  FAILED = 'FAILED',
  FETCHING = 'FETCHING',
  FINISHED = 'FINISHED',
  GENERATING = 'GENERATING',
  STOPPED = 'STOPPED',
}

export enum ChartType {
  AREA = 'AREA',
  BAR = 'BAR',
  GROUPED_BAR = 'GROUPED_BAR',
  LINE = 'LINE',
  MULTI_LINE = 'MULTI_LINE',
  PIE = 'PIE',
  STACKED_BAR = 'STACKED_BAR',
}
// ...
export type DetailStep = {
  __typename?: 'DetailStep';
  cteName?: Maybe<Scalars['String']>;
  sql: Scalars['String'];
  summary: Scalars['String'];
};
// ...
export type Error = {
  __typename?: 'Error';
  code?: Maybe<Scalars['String']>;
  message?: Maybe<Scalars['String']>;
  shortMessage?: Maybe<Scalars['String']>;
  stacktrace?: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type RecommendedQuestionsTask = {
  __typename?: 'RecommendedQuestionsTask';
  error?: Maybe<Error>;
  questions: Array<ResultQuestion>;
  status: RecommendedQuestionsTaskStatus;
};

export enum RecommendedQuestionsTaskStatus {
  FAILED = 'FAILED',
  FINISHED = 'FINISHED',
  GENERATING = 'GENERATING',
  NOT_STARTED = 'NOT_STARTED',
}
// ...
export type ResultCandidate = {
  __typename?: 'ResultCandidate';
  sql: Scalars['String'];
  sqlPair?: Maybe<SqlPair>;
  type: ResultCandidateType;
  view?: Maybe<ViewInfo>;
};

export enum ResultCandidateType {
  LLM = 'LLM',
  SQL_PAIR = 'SQL_PAIR',
  VIEW = 'VIEW',
}

export type ResultQuestion = {
  __typename?: 'ResultQuestion';
  category: Scalars['String'];
  question: Scalars['String'];
  sql: Scalars['String'];
};
// ...
export type SqlPair = {
  __typename?: 'SqlPair';
  createdAt?: Maybe<Scalars['String']>;
  id: Scalars['Int'];
  question: Scalars['String'];
  sql: Scalars['String'];
  updatedAt?: Maybe<Scalars['String']>;
};
// ...
export type ThreadResponse = {
  __typename?: 'ThreadResponse';
  adjustment?: Maybe<ThreadResponseAdjustment>;
  adjustmentTask?: Maybe<AdjustmentTask>;
  answerDetail?: Maybe<ThreadResponseAnswerDetail>;
  askingTask?: Maybe<AskingTask>;
  breakdownDetail?: Maybe<ThreadResponseBreakdownDetail>;
  chartDetail?: Maybe<ThreadResponseChartDetail>;
  id: Scalars['Int'];
  question: Scalars['String'];
  sql?: Maybe<Scalars['String']>;
  threadId: Scalars['Int'];
  view?: Maybe<ViewInfo>;
};

export type ThreadResponseAdjustment = {
  __typename?: 'ThreadResponseAdjustment';
  payload?: Maybe<Scalars['JSON']>;
  type: ThreadResponseAdjustmentType;
};

export enum ThreadResponseAdjustmentType {
  APPLY_SQL = 'APPLY_SQL',
  REASONING = 'REASONING',
}

export type ThreadResponseAnswerDetail = {
  __typename?: 'ThreadResponseAnswerDetail';
  content?: Maybe<Scalars['String']>;
  error?: Maybe<Error>;
  numRowsUsedInLLM?: Maybe<Scalars['Int']>;
  queryId?: Maybe<Scalars['String']>;
  status?: Maybe<ThreadResponseAnswerStatus>;
};

export enum ThreadResponseAnswerStatus {
  FAILED = 'FAILED',
  FETCHING_DATA = 'FETCHING_DATA',
  FINISHED = 'FINISHED',
  INTERRUPTED = 'INTERRUPTED',
  NOT_STARTED = 'NOT_STARTED',
  PREPROCESSING = 'PREPROCESSING',
  STREAMING = 'STREAMING',
}

export type ThreadResponseBreakdownDetail = {
  __typename?: 'ThreadResponseBreakdownDetail';
  description?: Maybe<Scalars['String']>;
  error?: Maybe<Error>;
  queryId?: Maybe<Scalars['String']>;
  status: AskingTaskStatus;
  steps?: Maybe<Array<DetailStep>>;
};

export type ThreadResponseChartDetail = {
  __typename?: 'ThreadResponseChartDetail';
  adjustment?: Maybe<Scalars['Boolean']>;
  canonicalizationVersion?: Maybe<Scalars['String']>;
  chartDataProfile?: Maybe<Scalars['JSON']>;
  chartSchema?: Maybe<Scalars['JSON']>;
  chartType?: Maybe<ChartType>;
  description?: Maybe<Scalars['String']>;
  error?: Maybe<Error>;
  queryId?: Maybe<Scalars['String']>;
  rawChartSchema?: Maybe<Scalars['JSON']>;
  renderHints?: Maybe<Scalars['JSON']>;
  status: ChartTaskStatus;
  validationErrors?: Maybe<Array<Scalars['String']>>;
};
// ...
export type ViewInfo = {
  __typename?: 'ViewInfo';
  displayName: Scalars['String'];
  id: Scalars['Int'];
  name: Scalars['String'];
  statement: Scalars['String'];
};